pan/layout: Use uint64_t types for offsets and sizes
As things are now, the pan_image_slice_layout::{offset,size}_B calculation
can exceed UINT32_MAX, and we silently droppin the upper 32-bit on the
floor. Same goes for the crc offset. Let's use a uint64_t to make sure
that doesn't happen.

While at it, move the two 64-bit field next to each other to avoid
padding, and document what size_B encodes.

It's hard to find the commit that introduced the problem with all the
code motion that happened this this struct was introduced, so I'll
just flag for backport.

Backport-to: 25.1
Tested-by: Eric R. Smith <eric.smith@collabora.com>
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/35555>

+++ src/panfrost/lib/pan_layout.c
@@ -120,7 +120,7 @@ pan_image_renderblock_size_el(uint64_t modifier, enum pipe_format format,
 
 static void
 init_slice_crc_info(unsigned arch, struct pan_image_slice_layout *slice,
-                    unsigned width_px, unsigned height_px, unsigned offset_B)
+                    unsigned width_px, unsigned height_px, uint64_t offset_B)
 {
    unsigned checksum_region_size_px = pan_meta_tile_size(arch);
    unsigned checksum_x_tile_per_region =
